backend/pyproject.toml
[project]
name = "rfp-ai-platform"
version = "0.1.0"
description = "AI-assisted RFP proposal PPT generation platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "langchain-community>=0.3.0",
  "langchain-chroma>=0.1.0",
  "chromadb>=0.5.0",
  "neo4j>=5.23.0",
  "tavily-python>=0.5.0",
  "python-pptx>=1.0.2",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
  "celery>=5.4.0",
  "httpx>=0.27.0",
  "tenacity>=9.0.0",
  "orjson>=3.10.0",
  "structlog>=24.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
